Update README and improve project implementation

+++ src/main/java/com/utown/utown_backend/dto/request/OptionRequestDTO.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -18,5 +20,5 @@ public class OptionRequestDTO {
     private String name;
 
     @NotNull(message = "Extra Price is required")
-    private Double extraPrice;
+    private BigDecimal extraPrice;
 }
