+++ src/main/java/com/utown/utown_backend/dto/response/OptionResponseDTO.java
@@ -2,6 +2,8 @@ package com.utown.utown_backend.dto.response;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 @Builder
@@ -9,7 +11,7 @@ public class OptionResponseDTO {
 
     private final long id;
     private final String name;
-    private final Double extraPrice;
+    private final BigDecimal extraPrice;
     private final Long dishId;
     private final String dishName;
 }
